// src/Features/RegisterPatientForm/Model/consts/consts.ts
// for enums consts wcich compiles in object

export const enum  ValidateCreatePatientError {
  INCORRECT_FIRST_NAME = "INCORRECT_FIRST_NAME",
  INCORRECT_LAST_NAME = "INCORRECT_LAST_NAME",
  INCORRECT_PATRONYMIC = "INCORRECT_PATRONYMIC",
  INCORRECT_LOGIN = "INCORRECT_LOGIN",
  INCORRECT_DATE_BIRTHDAY="INCORRECT_DATE_BIRTHDAY",
  INCORRECT_ADDRESS="INCORRECT_ADDRESS",
  INCORRECT_PHONE_NUMBER="INCORRECT_PHONE_NUMBER",
  INCORRECT_PASSWORD="INCORRECT_PASSWORD",
  INCORRECT_PASSPORT="INCORRECT_PASSPORT",
  NO_DATA = "NO_DATA",
  SERVER_ERROR = "SERVER_ERROR",
}